token revocation

+++ backend/opinions-sv/service/src/main/kotlin/com/r8n/backend/opinions/service/OpinionService.kt
@@ -4,6 +4,7 @@ import com.r8n.backend.opinions.domain.Opinion
 import com.r8n.backend.opinions.domain.OpinionStatusEnum
 import com.r8n.backend.opinions.persistence.OpinionPersistence
 import com.r8n.backend.opinions.provider.database.OpinionRepository
+import com.r8n.backend.security.CurrentUserIdentifier.getCurrentUserId
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -20,19 +21,26 @@ class OpinionService(
 ) {
     fun getOpinion(id: UUID): Opinion {
         val opinion = opinionRepository.findById(id).orElseThrow { ResponseStatusException(HttpStatus.NOT_FOUND) }
+        if (!getCurrentUserId().isOwnerOf(opinion)) {
+            throw ResponseStatusException(HttpStatus.FORBIDDEN)
+        }
         return opinion.toDomain()
     }
+
+    private fun UUID.isOwnerOf(opinion: OpinionPersistence): Boolean = this == opinion.owner
 
     fun getOpinionFor(subjectId: UUID): Opinion {
         val opinion =
             opinionRepository.findFirstBySubjectOrderByTimestampDesc(subjectId)
                 ?: throw ResponseStatusException(HttpStatus.NOT_FOUND)
+        if (opinion.owner != getCurrentUserId()) {
+            throw ResponseStatusException(HttpStatus.FORBIDDEN)
+        }
         return opinion.toDomain()
     }
 
     @Transactional
     fun createOpinion(
-        ownerId: UUID,
         subjectId: UUID,
         subjective: List<String>,
         objective: List<String>,
@@ -41,7 +49,7 @@ class OpinionService(
         val opinion =
             opinionRepository.save(
                 OpinionPersistence(
-                    owner = ownerId,
+                    owner = getCurrentUserId(),
                     subject = subjectId,
                     mark = mark,
                     status = OpinionStatusEnum.DRAFT,
@@ -64,6 +72,9 @@ class OpinionService(
                 .findById(
                     opinionId,
                 ).orElseThrow { ResponseStatusException(HttpStatus.NOT_FOUND) }
+        if (opinion.owner != getCurrentUserId()) {
+            throw ResponseStatusException(HttpStatus.FORBIDDEN)
+        }
         opinion.mark = mark
         opinion.timestamp = Instant.now()
         val savedOpinion = opinionRepository.save(opinion)
@@ -78,6 +89,9 @@ class OpinionService(
                 .findById(
                     opinionId,
                 ).orElseThrow { ResponseStatusException(HttpStatus.NOT_FOUND) }
+        if (opinion.owner != getCurrentUserId()) {
+            throw ResponseStatusException(HttpStatus.FORBIDDEN)
+        }
         opinionRepository.delete(opinion)
     }
